Email Methods implemented
