+++ src/sample/SendEmail.java
@@ -2,48 +2,38 @@ package sample;
 import java.util.*;
 import javax.mail.*;
 import javax.mail.internet.*;
-import javax.activation.*;
+import java.util.Properties;
 
-public class SendEmail {public static void main(String [] args) {
-    // Recipient's email ID needs to be mentioned.
-    String to = "stiv.hkr@gmail.com";
+public class SendEmail {
 
-    // Sender's email ID needs to be mentioned
-    String from = "stivo1999@gmail.com";
+    public static void send(String to,String sub,String msg){
+        //Get properties object
+        Properties props = new Properties();
+        props.put("mail.smtp.host", "smtp.gmail.com");
+        props.put("mail.smtp.socketFactory.port", "465");
+        props.put("mail.smtp.socketFactory.class",
+                "javax.net.ssl.SSLSocketFactory");
+        props.put("mail.smtp.auth", "true");
+        props.put("mail.smtp.port", "465");
+        //get Session
+        Session session = Session.getDefaultInstance(props,
+                new javax.mail.Authenticator() {
+                    protected PasswordAuthentication getPasswordAuthentication() {
+                        return new PasswordAuthentication("sajj.hkr.pc2@gmail.com","AJY36e9fvw26in9");
+                    }
+                });
+        //compose message
+        try {
+            MimeMessage message = new MimeMessage(session);
+            message.addRecipient(Message.RecipientType.TO,new InternetAddress(to));
+            message.setSubject(sub);
+            message.setText(msg);
+            //send message
+            Transport.send(message);
+            System.out.println("message sent successfully");
+        } catch (MessagingException e) {throw new RuntimeException(e);}
 
-    // Assuming you are sending email from localhost
-    String host = "localhost";
-
-    // Get system properties
-    Properties properties = System.getProperties();
-
-    // Setup mail server
-    properties.setProperty("mail.smtp.host", host);
-
-    // Get the default Session object.
-    Session session = Session.getDefaultInstance(properties);
-
-    try {
-        // Create a default MimeMessage object.
-        MimeMessage message = new MimeMessage(session);
-
-        // Set From: header field of the header.
-        message.setFrom(new InternetAddress(from));
-
-        // Set To: header field of the header.
-        message.addRecipient(Message.RecipientType.TO, new InternetAddress(to));
-
-        // Set Subject: header field
-        message.setSubject("This is the Subject Line!");
-
-        // Now set the actual message
-        message.setText("This is actual message");
-
-        // Send message
-        Transport.send(message);
-        System.out.println("Sent message successfully....");
-    } catch (MessagingException mex) {
-        mex.printStackTrace();
     }
 }
-}
+
+
