--- src/sample/SendEmailStivs.java
@@ -1,49 +1,0 @@
-package sample;
-import java.util.*;
-import javax.mail.*;
-import javax.mail.internet.*;
-import java.util.Properties;
-
-public class SendEmailStivs {
-
-    /*    public void SendEmail(){
-        Properties props = new Properties();
-        props.put("mail.smtp.host","smtp.gmail.gom");
-        props.put("mail.smtp.socketFactory.port","465");
-        props.put("mail.smtp.socketFactory.class","javax.net.ssl.SSLSocketFactory");
-        props.put("mail.smtp.auth","true");
-        props.put("mail.smtp.port","465");
-        Session
-        }
-    }*/
-    public static void send(String from,String password,String to,String sub,String msg){
-        //Get properties object
-        Properties props = new Properties();
-        props.put("mail.smtp.host", "smtp.gmail.com");
-        props.put("mail.smtp.socketFactory.port", "465");
-        props.put("mail.smtp.socketFactory.class",
-                "javax.net.ssl.SSLSocketFactory");
-        props.put("mail.smtp.auth", "true");
-        props.put("mail.smtp.port", "465");
-        //get Session
-        Session session = Session.getDefaultInstance(props,
-                new javax.mail.Authenticator() {
-                    protected PasswordAuthentication getPasswordAuthentication() {
-                        return new PasswordAuthentication(from,password);
-                    }
-                });
-        //compose message
-        try {
-            MimeMessage message = new MimeMessage(session);
-            message.addRecipient(Message.RecipientType.TO,new InternetAddress(to));
-            message.setSubject(sub);
-            message.setText(msg);
-            //send message
-            Transport.send(message);
-            System.out.println("message sent successfully");
-        } catch (MessagingException e) {throw new RuntimeException(e);}
-
-    }
-}
-
-
